Breaking code into components
